requirements.txt
flask
requests
fastapi
pydantic
python-dotenv
psycopg2-binary
uvicorn
joblib
pandas
numpy
scikit-learn
